scripts/requirements.txt
airtable-python-wrapper==0.15.3
openai>=1.0.0
python-dotenv==1.0.0
tenacity==8.2.3
requests==2.31.0